kronos: Replace RSS flash with SI NVM flash
As the maximum size of RSS flash is 8MB which is not enough
if we add multiple images of SI to the flash on rdkronos
platform, so replace it with the safety island NVM flash.

Add RSS_SI_FLASH macro to use the SI NVM flash in common code.

Replace the old FVP parameter to the new one as below when using
SI NVM flash.
old:
--data css.sysctrl.rss.cpu=<PATH-TO-BINARY>/rss-flash-image-fvp-rd-kronos.wic.nopt@0xB0000000
new:
--parameter css.sysctrl.si.nvm_loader.fname=<PATH-TO-BINARY>/rss-flash-image-fvp-rd-kronos.wic.nopt

Upstream-Status: Pending
Signed-off-by: Jun Zhang <jun.zhang@arm.com>