platform/rss/common: Introduce NI-710-AE APU driver
The NCI interconnect in the Safety Island is the NI-710-AE, and
the APU driver of NI-710-AE is similar as the one for NCI-Tower.
Since there is an ongoing work of refactoring the NCI-Tower, to
save effort of rebasing the code, add an NI-710-AE APU driver by
duplicating the key functions of APU driver from the NCI-Tower.

Upstream-Status: Pending
Signed-off-by: Henry Wang <Henry.Wang@arm.com>
Signed-off-by: Wei Chen <wei.chen@arm.com>