RSS: Make maximum number of concurrent requests configurable
The maximum number of concurrent requests to RSS is hardcoded
to two. This value needs to be configurable as there are
four connections which may be established with
the RSS at the same time in Kronos:
1- Application Processor
2- Safety Island Cluster 0, 1 and 2.
Make this value configurable and set it to four in kronos so
none of the connections get a connection busy reply.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>