rss/kronos: Improvements for asni_rss_mm APU configs
Make the runtime asni_rss_mm APU config background to avoid conflicts
when changing the boot time and runtime configs.

Also, although there is no functional difference, it is still better
to use consecutive region numbers when programming the APU, take
the opportunity to fix in this commit.

Upstream-Status: Pending
Signed-off-by: Henry Wang <Henry.Wang@arm.com>