platform: Disable interrupts in stdio output string function
UART stdout is not designed to be reentrant.
If there is an ongoing stdio output operation and an interrupt
occurs which also tries to use stdio output, this causes
an undesirable effect of either interleaved messages or even
printing garbage data infinitely.

Disable the interrupts on entering stdio_output_string and
re-enable it before returning.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>