rss/kronos: Add FVP shutdown handler
In Kronos, There is no differentiation between a reset and
a shutdown request. A shutdown request results in a system
reset which is an incorrect behavior.

This is because the FVP model can not be turned off through
software.

Create a work around by adding a shutdown handler that
prints a shutdown message, this message can be detected
by the FVP shutdown_tag parameter causing the FVP to
terminate.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>